I accidently messed up the ortholog generator.
I have a comment asking why I made a specific query.  Here is the
comment:
"I am not sure what I was doing with the following 5 lines of code."

Well, those lines were to get the set of species for which to collect
the orthologs.  Unfortunately, I immediately thereafter did not use that
set of species and so did not actually download the orthologs.

I fixed this and will remove the comment now.