playing with a fungidb entry.
While at it, I cleaned out some of the extraneous parameters in the
webservices functions.  A few functions like in the post_eupath family
had a series of redundant input parameters.  I removed them, which means
that I will no longer be able to call arbitrary functions with no
inputs, but so what.