Clearing shenanigans caused by matching against GenomeInfoDb.
I should have noticed, but using GenomeInfoDB to set the species ID for
the various strains and such from the eupathdb is both excellent and
terrible.  Excellent because it makes many things more canonical,
terrible because it makes different strains have the same ID, and thus
causes some invocations to fail.

In any event, I fixed that peculiarity by adding a parameter to the
various functions which query the species.  I need to update the
docstrings still, though.