a larger change than intended
All I wanted to do was change the news file and understand taxonomy ids
a little better.

Instead I rewrote the logic for handling taxonomy ids and resolving
conflicts/NAs between the eupathdb taxonomy ids vs. genomeInfoDB's ids.
Along the way towards doing that, I changed the make-data.R script so
that it should theoretically both create all the packages and put the s3
files in place for annotationHub.  Along the way for that, I found that
the GO/pathway/reactome ids for organismdbi were not being properly set.

So the end result should be a nice improvement which should allow future
annotationhub releases to be much easier.  Though having said that, the
organismdbi logic is still confusing vis a vis where it puts its files.