Initial code to download all SNPs from the EuPathDB.
This might be a bit invasive on my part, so before actually adding this
to anything, I am going to write Cristina and make sure they do not mind
me doing this.  However, I think it is worth committing because it
provies an example of how one might invoke a more complex REST search.