First time working on this in a while.
Simplified portions of it, working toward generating packages/sqlite
files from within the tests in the hopes that it will make it easier to
chase down problems.  Thus there is a new test file 040_toxodb, which
currently builds the bsgenome and orgdb files.  This leaves the
granges/txdb and organismdbi files to step through.